+++ lib/geocoder/lookups/google_places_details.rb
@@ -19,14 +19,21 @@ module Geocoder
       private
 
       def base_query_url(query)
-        "#{protocol}://maps.googleapis.com/maps/api/place/details/json?"
+        use_new_places_api = query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+        if use_new_places_api
+          "#{protocol}://places.googleapis.com/v1/places/#{query.text}?"
+        else
+          "#{protocol}://maps.googleapis.com/maps/api/place/details/json?"
+        end
       end
 
       def result_root_attr
-        'result'
+        use_new_places_api = @query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+        use_new_places_api ? nil : 'result'
       end
 
       def results(query)
+        @query = query
         result = super(query)
         return [result] unless result.is_a? Array
 
@@ -53,11 +60,31 @@ module Geocoder
       end
 
       def query_url_google_params(query)
-        {
-          placeid: query.text,
-          fields: fields(query),
-          language: query.language || configuration.language
-        }
+        use_new_places_api = query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+
+        if use_new_places_api
+          {
+            fields: fields(query),
+            languageCode: query.language || configuration.language
+          }
+        else
+          {
+            placeid: query.text,
+            fields: fields(query),
+            language: query.language || configuration.language
+          }
+        end
+      end
+
+      def parse_raw_data(raw_data)
+        use_new_places_api = @query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+
+        if use_new_places_api
+          # For the new API, the top-level field is the place object itself
+          super(raw_data)
+        else
+          super(raw_data)
+        end
       end
     end
   end