+++ lib/geocoder/lookups/google_places_search.rb
@@ -19,21 +19,39 @@ module Geocoder
       private
 
       def result_root_attr
-        'candidates'
+        use_new_places_api = @query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+        use_new_places_api ? 'places' : 'candidates'
       end
 
       def base_query_url(query)
-        "#{protocol}://maps.googleapis.com/maps/api/place/findplacefromtext/json?"
+        use_new_places_api = query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+        if use_new_places_api
+          "#{protocol}://places.googleapis.com/v1/places:searchText?"
+        else
+          "#{protocol}://maps.googleapis.com/maps/api/place/findplacefromtext/json?"
+        end
       end
 
       def query_url_google_params(query)
-        {
-          input: query.text,
-          inputtype: 'textquery',
-          fields: fields(query),
-          locationbias: locationbias(query),
-          language: query.language || configuration.language
-        }
+        @query = query
+        use_new_places_api = query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+
+        if use_new_places_api
+          {
+            textQuery: query.text,
+            fields: fields(query),
+            locationBias: locationbias(query),
+            languageCode: query.language || configuration.language
+          }
+        else
+          {
+            input: query.text,
+            inputtype: 'textquery',
+            fields: fields(query),
+            locationbias: locationbias(query),
+            language: query.language || configuration.language
+          }
+        end
       end
 
       def fields(query)
@@ -49,10 +67,18 @@ module Geocoder
       end
 
       def default_fields
-        basic = %w[business_status formatted_address geometry icon name 
-          photos place_id plus_code types]
-        contact = %w[opening_hours]
-        atmosphere = %W[price_level rating user_ratings_total]
+        use_new_places_api = @query.options.fetch(:use_new_places_api, configuration.use_new_places_api)
+
+        if use_new_places_api
+          basic = %w[businessStatus formattedAddress location viewport iconMaskBaseUri iconBackgroundColor displayName photos id plusCode types]
+          contact = %w[regularOpeningHours]
+          atmosphere = %W[priceLevel rating userRatingsTotal]
+        else
+          basic = %w[business_status formatted_address geometry icon name photos place_id plus_code types]
+          contact = %w[opening_hours]
+          atmosphere = %W[price_level rating user_ratings_total]
+        end
+
         format_fields(basic, contact, atmosphere)
       end